+++ lib/image_picker_web.dart
@@ -1,22 +1,24 @@
 library image_picker_web;
+export 'src/Models/Types.dart';
 
 import 'dart:convert';
-import 'dart:typed_data';
 
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
 
 import 'src/web_image_picker.dart';
-import 'src/bean/web_media_info.dart';
+import 'src/Models/Types.dart';
 
 class ImagePickerWeb {
   static void registerWith(Registrar registrar) {
-    final channel = MethodChannel('image_picker_web',
-        const StandardMethodCodec(), registrar.messenger);
+    final channel = MethodChannel(
+        'image_picker_web', const StandardMethodCodec(), registrar.messenger);
     final instance = WebImagePicker();
     channel.setMethodCallHandler((call) async {
       switch (call.method) {
+        case 'pickFile':
+          return await instance.pickFile(call.arguments);
         case 'pickImage':
           return await instance.pickImage();
         case 'pickVideo':
@@ -30,20 +32,37 @@ class ImagePickerWeb {
   static const MethodChannel _methodChannel =
       const MethodChannel('image_picker_web');
 
-  static Future<dynamic> getImage({bool asUint8List = false}) async {
-    final data =
-        await _methodChannel.invokeMapMethod<String, dynamic>('pickImage');
-    final imageName = data['name'];
-    final imageData = base64.decode(data['data']);
-    if (asUint8List) {
-      return imageData;
+  static Future<dynamic> getImage({@required ImageType outputType}) async {
+    if (!(outputType is ImageType)) {
+      throw ArgumentError('outputType has to be from Type: ImageType if you call getImage()');
     }
-    return Image.memory(imageData, semanticLabel: imageName);
+    switch (outputType) {
+      case ImageType.file:
+        return await _methodChannel
+            .invokeMapMethod<String, dynamic>('pickFile', ['image']);
+        break;
+      case ImageType.bytes:
+        final data =
+            await _methodChannel.invokeMapMethod<String, dynamic>('pickImage');
+        final imageData = base64.decode(data['data']);
+        return imageData;
+        break;
+      case ImageType.widget:
+        final data =
+            await _methodChannel.invokeMapMethod<String, dynamic>('pickImage');
+        final imageName = data['name'];
+        final imageData = base64.decode(data['data']);
+        return Image.memory(imageData, semanticLabel: imageName);
+        break;
+      default:
+        return null;
+        break;
+    }
   }
 
   static Future<MediaInfo> get getImageInfo async {
     final data =
-    await _methodChannel.invokeMapMethod<String, dynamic>('pickImage');
+        await _methodChannel.invokeMapMethod<String, dynamic>('pickImage');
     MediaInfo _webImageInfo = MediaInfo();
     _webImageInfo.fileName = data['name'];
     _webImageInfo.filePath = data['path'];
@@ -53,16 +72,30 @@ class ImagePickerWeb {
     return _webImageInfo;
   }
 
-  static Future<Uint8List> get getVideo async {
-    final data =
-        await _methodChannel.invokeMapMethod<String, dynamic>('pickVideo');
-    final videoData = base64.decode(data['data']);
-    return videoData;
+  static Future<dynamic> getVideo({@required VideoType outputType}) async {
+    if (!(outputType is VideoType)) {
+      throw ArgumentError('outputType has to be from Type: VideoType if you call getVideo()');
+    }
+    switch (outputType) {
+      case VideoType.file:
+        return await _methodChannel
+            .invokeMapMethod<String, dynamic>('pickFile', ['video']);
+        break;
+      case VideoType.bytes:
+        final data =
+            await _methodChannel.invokeMapMethod<String, dynamic>('pickVideo');
+        final imageData = base64.decode(data['data']);
+        return imageData;
+        break;
+      default:
+        return null;
+        break;
+    }
   }
 
   static Future<MediaInfo> get getVideoInfo async {
     final data =
-    await _methodChannel.invokeMapMethod<String, dynamic>('pickVideo');
+        await _methodChannel.invokeMapMethod<String, dynamic>('pickVideo');
     MediaInfo _webVideoInfo = MediaInfo();
     _webVideoInfo.fileName = data['name'];
     _webVideoInfo.filePath = data['path'];