+++ lib/src/Models/Types.dart
@@ -1,9 +1,20 @@
 import 'dart:typed_data';
 
-class MediaInfo{
+class MediaInfo {
   String fileName;
   String filePath;
   String base64;
   String base64WithScheme;
   Uint8List data;
+}
+
+enum ImageType {
+  file,
+  bytes,
+  widget
+}
+
+enum VideoType {
+  file,
+  bytes
 }